Move useMediumZoom to blog post hooks and remove all zoom elements when cleaning up
