+++ src/components/Blog/BlogPost/index.tsx
@@ -1,12 +1,11 @@
-import React, { useLayoutEffect, useRef, RefObject } from 'react';
+import React, { useEffect, useRef, RefObject } from 'react';
 import shortid from 'shortid';
-import mediumZoom from 'medium-zoom';
 
 import Page from '../../Layout/Page';
 import Comments from '../Comments';
 import styles from './BlogPost.module.css';
 import BlogPostNewsletterSignup from '../BlogPostNewsletterSignup';
-import { prefersDarkMode } from '../../../utils/helpers';
+import { useMediumZoom } from './BlogPostHooks';
 
 const addKeys = (arr = [], propName: string) =>
   arr.map(item => ({ key: shortid.generate(), [propName]: item }));
@@ -17,28 +16,6 @@ interface IBlogPostProps {
   body: any;
   tags: any;
 }
-
-const useMediumZoom = (postBodyRef: RefObject<HTMLDivElement>) => {
-  let zoom;
-
-  useLayoutEffect(() => {
-    if (postBodyRef) {
-      const $postBody = postBodyRef.current as HTMLDivElement;
-      const $images = $postBody ? $postBody.querySelectorAll('img') : [];
-
-      if ($images.length) {
-        zoom = mediumZoom($images, {
-          background: prefersDarkMode()
-            ? 'rgba(0, 0, 0, .9)'
-            : 'rgba(255,255,255, .83)',
-          scrollOffset: 10
-        });
-      }
-    }
-
-    () => zoom && zoom.detach();
-  }, [postBodyRef]);
-};
 
 const BlogPost: React.FC<IBlogPostProps> = ({ title, date, body, tags }) => {
   const postBodyRef = useRef();
